Remove the MeasurableUnitEntity mixin
